Organizing code in OOP requirements and Adding explaination in readme file
