+++ README.md
@@ -7,3 +7,23 @@ This repository contains the implementation of a Convolutional Neural Network (C
 ## Introduction
 The provided code implements a custom CNN architecture with rotational and scale equivariant convolutional layers. These layers aim to capture rotational and scale-invariant features in the input data, making the model suitable for a variety of image recognition tasks.
 
+## Running the Code
+To run the code:
+1. Create python vitual environment and activate it.
+    1.1 For Windows:
+    cd path/to/your/project
+    python3 -m venv venv_name
+    .\venv_name\Scripts\activate
+
+    1.2 For macOS/Linux:
+    cd path/to/your/project
+    python3 -m venv venv_name
+    source venv_name/bin/activate
+
+2. Install dependencies
+    pip install -r requirements.txt
+
+3. Run the main.py file
+    python3 main.py
+
+
