+++ requirements.txt
@@ -1,0 +1,6 @@
+# Usage: pip install -r requirements.txt
+
+matplotlib==3.8.2
+numpy==1.26.3
+torch==2.1.2
+torchvision==0.16.2
